refactored ProfileScreen
